Refactor namespaces from Goods to Masters.
Updated various files to rename namespaces from `Goods\Catalogue` and `Goods\MasterAsset` to `Masters` for better organization and consistency. Adjusted references in affected classes, actions, tests, and enums accordingly.
