+++ app/Transfers/Aurora/FetchAuroraMasterAsset.php
@@ -10,7 +10,7 @@ namespace App\Transfers\Aurora;
 
 use App\Actions\Helpers\CurrencyExchange\GetCurrencyExchange;
 use App\Enums\Catalogue\Product\ProductUnitRelationshipType;
-use App\Enums\Goods\MasterAsset\MasterAssetTypeEnum;
+use App\Enums\Masters\MasterAsset\MasterAssetTypeEnum;
 use App\Models\Masters\MasterProductCategory;
 use Illuminate\Support\Facades\DB;
